# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Memo {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  fileName?: string | null;
  createdAt: string;
  summaryCount?: number;
}

export interface MemoInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  fileName?: string;
}

export interface ExtractedText {
  text: string;
  fileName: string;
}

export interface Summary {
  id: number;
  memoId: number;
  audience: string;
  goal: string;
  format: string;
  content: string;
  createdAt: string;
}

/**
 * Output format
 */
export type SummaryRequestFormat =
  (typeof SummaryRequestFormat)[keyof typeof SummaryRequestFormat];

export const SummaryRequestFormat = {
  email: "email",
  "one-pager": "one-pager",
  slack: "slack",
} as const;

export interface SummaryRequest {
  /** Who will receive this summary (e.g. in-house team, board, business unit) */
  audience: string;
  /** What they want from the advice (e.g. understand key risks, decide on next steps) */
  goal: string;
  /** Output format */
  format: SummaryRequestFormat;
}

export interface MemoStats {
  totalMemos: number;
  totalSummaries: number;
  recentMemos: Memo[];
}