# lib/api-zod/src/generated/types/extractedText.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedText {
  text: string;
  fileName: string;
}