# lib/api-zod/src/generated/types/memo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */

export interface Memo {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  fileName?: string | null;
  createdAt: Date;
  summaryCount?: number;
}