# lib/api-zod/src/generated/types/memoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */

export interface MemoInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  fileName?: string;
}