# lib/api-zod/src/generated/types/memoStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */
import type { Memo } from "./memo";

export interface MemoStats {
  totalMemos: number;
  totalSummaries: number;
  recentMemos: Memo[];
}