# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */

export interface Summary {
  id: number;
  memoId: number;
  audience: string;
  goal: string;
  format: string;
  content: string;
  createdAt: Date;
}