# lib/api-zod/src/generated/types/summaryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Memo Simplifier API
 * OpenAPI spec version: 0.1.0
 */
import type { SummaryRequestFormat } from "./summaryRequestFormat";

export interface SummaryRequest {
  /** Who will receive this summary (e.g. in-house team, board, business unit) */
  audience: string;
  /** What they want from the advice (e.g. understand key risks, decide on next steps) */
  goal: string;
  /** Output format */
  format: SummaryRequestFormat;
}